+++ demo/test/components/previews/primer/docs/counter_component_preview.rb
@@ -1,9 +1,0 @@
-module Primer
-  module Docs
-    class CounterComponentPreview < ViewComponent::Preview
-      def default; end
-
-      def schemes; end
-    end
-  end
-end