+++ demo/test/components/previews/primer/docs/dropdown_component_preview.rb
@@ -1,9 +1,0 @@
-module Primer
-  module Docs
-    class DropdownComponentPreview < ViewComponent::Preview
-      def default; end
-
-      def with_direction; end
-    end
-  end
-end