+++ demo/test/components/previews/primer/docs/flex_item_component_preview.rb
@@ -1,7 +1,0 @@
-module Primer
-  module Docs
-    class FlexItemComponentPreview < ViewComponent::Preview
-      def default; end
-    end
-  end
-end