+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 
 ### Misc
 
+* Don't commit auto-generated component previews.
+
+### Misc
+
 * Provide linters for component migrations.
 
     *Manuel Puyol*